実装切り出し
